change to unsigned long for comparison with PixelArray

+++ src_c/surface.c
@@ -1014,7 +1014,7 @@ surf_map_rgb(PyObject *self, PyObject *args)
         return RAISE(pgExc_SDLError, "display Surface quit");
 
     color = SDL_MapRGBA(surf->format, rgba[0], rgba[1], rgba[2], rgba[3]);
-    return PyInt_FromLong(color);
+    return PyInt_FromLong((unsigned int) color);
 }
 
 static PyObject *
